✨ #98 Finish custom cycle keybind logic
